Fixed php warnings.
git-svn-id: svn://svn.ispconfig.org/ispconfig3/branches/ispconfig-3.0.3@2121 c1a6f2c7-8504-0410-933a-a4dae0184b0a

+++ server/plugins-available/apache2_plugin.inc.php
@@ -915,7 +915,7 @@ class apache2_plugin {
 		}
 		
 		// Remove the backup copy of the config file.
-		unlink($vhost_file.'~');
+		if(@is_file($vhost_file.'~')) unlink($vhost_file.'~');
 		
 
 		//* Unset action to clean it for next processed vhost.
